+++ scribbler/templatetags/scribbler_tags.py
@@ -4,10 +4,11 @@ from __future__ import unicode_literals
 from django import template
 from django.conf import settings
 from django.core.cache import cache
+from django.contrib.contenttypes.generic import ContentType
 from django.core.exceptions import ImproperlyConfigured
 
 from scribbler.conf import CACHE_TIMEOUT, CACHE_KEY_FUNCTION
-from scribbler.forms import ScribbleForm
+from scribbler.forms import ScribbleForm, FieldScribbleForm
 from scribbler.models import Scribble
 from scribbler.views import build_scribble_context
 
@@ -17,7 +18,7 @@ register = template.Library()
 
 class ScribbleNode(template.Node):
     "Render snippet or default block content."
-    
+
     child_nodelists = ('nodelist_default', )
 
     def __init__(self, slug, nodelist, raw):
@@ -52,7 +53,7 @@ class ScribbleNode(template.Node):
         scribble_context = build_scribble_context(scribble, request)
         content = scribble_template.render(scribble_context)
         wrapper_template = template.loader.get_template('scribbler/scribble-wrapper.html')
-        context['scribble'] = scribble        
+        context['scribble'] = scribble
         context['rendered_scribble'] = content
         user = context.get('user', None)
         show_controls = False
@@ -98,7 +99,7 @@ def rebuild_template_string(tokens):
                 template.COMMENT_TAG_END,
             )
         result = '{0}{1}'.format(result, value)
-    return result     
+    return result
 
 
 @register.tag
@@ -125,3 +126,49 @@ def scribble(parser, token):
     parser.delete_first_token()
     raw = rebuild_template_string(tokens)
     return ScribbleNode(slug=slug, nodelist=nodelist, raw=raw)
+
+
+@register.simple_tag(takes_context=True)
+def scribble_field(context, model_instance, field_name):
+    """
+    Renders a scribble-able field from a model instance.
+
+    Usage:
+    {% scribble_field model_instance 'field_name' %}
+    """
+
+    # TODO: This should maybe be a utility funciton
+    request = context.get('request', None)
+    if request is None: # pragma: no cover
+        if settings.DEBUG:
+            msg = '"django.core.context_processors.request" is required to use django-scribbler'
+            raise ImproperlyConfigured(msg)
+        else:
+            return ''
+
+    model_content_type = ContentType.objects.get_for_model(model_instance)
+    field_value = getattr(model_instance, field_name)
+    scribble_template = template.Template(field_value)
+    scribble_context = build_scribble_context(None, request)
+    rendered_content = scribble_template.render(scribble_context)
+    context['rendered_scribble'] = rendered_content
+
+    user = context.get('user', None)
+    can_edit = False
+    if user:
+        perm_name = '{0}.change_{1}'.format(
+            model_content_type.app_label,
+            model_content_type.name
+        )
+        can_edit = user.has_perm(perm_name)
+    if can_edit:
+        context['scribble_form'] = FieldScribbleForm(
+            model_content_type, model_instance.pk, field_name, field_value=field_value)
+    context['show_controls'] = can_edit
+    context['can_add_scribble'] = False
+    context['can_edit_scribble'] = can_edit
+    context['can_delete_scribble'] = False
+    context['raw_content'] = field_value
+    wrapper_template = template.loader.get_template('scribbler/scribble-wrapper.html')
+    return wrapper_template.render(context)
+