+++ scribbler/urls.py
@@ -13,4 +13,6 @@ urlpatterns = patterns('scribbler.views',
     url('^create/$', 'create_edit_scribble', name='create-scribble'),
     url('^edit/(?P<scribble_id>(\d+))/$', 'create_edit_scribble', name='edit-scribble'),
     url('^delete/(?P<scribble_id>(\d+))/$', 'delete_scribble', name='delete-scribble'),
+    url('^edit-field/(?P<ct_pk>(\d+))/(?P<instance_pk>(\d+))/(?P<field_name>(\w+))/$',
+       'edit_scribble_field', name='edit-scribble-field'),
 )